Defines the start point for each day to simulate.

This table is a random sample of all trips in 2023, generated using the reservoir method of duckdb with a seed of 3518, as described in vignette("02-business-understanding").

Usage

SimulationStartDay

Format

SimulationStartDay

A data frame with 60 rows and 22 columns:

trip_id

An unique indentifier for each trip.

hvfhs_license_num

The TLC license number of the HVFHS base or business. See Details.

dispatching_base_num

The TLC Base License Number of the base that dispatched the trip.

originating_base_num

Base number of the base that received the original trip request.

request_datetime

Date/time when passenger requested to be picked up.

on_scene_datetime

Date/time when driver arrived at the pick-up location (Accessible Vehicles-only).

pickup_datetime

The date and time of the trip pick-up.

dropoff_datetime

The date and time of the trip drop-off.

PULocationID

TLC Taxi Zone in which the trip began.

DOLocationID

TLC Taxi Zone in which the trip ended.

trip_miles

Total miles for passenger trip.

trip_time

Total time in seconds for passenger trip.

base_passenger_fare

Base passenger fare before tolls, tips, taxes, and fees.

tolls

Total amount of all tolls paid in trip.

bcf

Total amount collected in trip for Black Car Fund.

sales_tax

Total amount collected in trip for NYS sales tax.

congestion_surcharge

Total amount collected in trip for NYS congestion surcharge.

airport_fee

$2.50 for both drop off and pick up at LaGuardia, Newark, and John F. Kennedy airports.

tips

Total amount of tips received from passenger.

driver_pay

Total driver pay (not including tolls or tips and net of commission, surcharges, or taxes).

shared_request_flag

Did the passenger agree to a shared/pooled ride? (Y/N)

shared_match_flag

Did the passenger share the vehicle with another passenger? (Y/N)

access_a_ride_flag

Was the trip administered on behalf of the Metropolitan Transportation Authority (MTA)? (Y/N)

wav_request_flag

Did the passenger request a wheelchair-accessible vehicle (WAV)? (Y/N)

wav_match_flag

Did the trip occur in a wheelchair-accessible vehicle (WAV)? (Y/N)

performance_per_hour

Represent the return if the would have taken an hour

Source

https://www.nyc.gov/site/tlc/about/tlc-trip-record-data.page