Counts rows based con columns provided

It counts the number rows where each unique value repeated in the columns selected arranging there results in descent order and adds a percentage column after collecting the results from arrow.

Usage

count_pct(x, ..., wt = NULL, sort = TRUE, digits = 3L)

Arguments

x

A lazy data frame arrow connection.

...

Variables to group by

wt

If a variable, computes sum(wt) for each group.

sort

If TRUE, will show the largest groups at the top.

digits

Integer indicating the number of decimal places

Value

A data.frame

Examples

set.seed(1234)

arrow_con <- 
  data.frame(char1 = sample(LETTERS[1:3], 100, replace = TRUE),
             char2 = sample(LETTERS[4:6], 100, replace = TRUE)) |>
  arrow::arrow_table()

# One var
count_pct(arrow_con, char1)
#>     char1     n   pct pct_cumulative
#>    <char> <int> <num>          <num>
#> 1:      B    40  0.40           0.40
#> 2:      C    31  0.31           0.71
#> 3:      A    29  0.29           1.00

# Two vars
dt1 <-count_pct(arrow_con, char1, char2)
dt1
#>     char1  char2     n   pct pct_cumulative
#>    <char> <char> <int> <num>          <num>
#> 1:      B      F    17  0.17           0.17
#> 2:      B      E    14  0.14           0.31
#> 3:      C      F    12  0.12           0.43
#> 4:      A      F    10  0.10           0.53
#> 5:      A      E    10  0.10           0.63
#> 6:      C      D    10  0.10           0.73
#> 7:      B      D     9  0.09           0.82
#> 8:      A      D     9  0.09           0.91
#> 9:      C      E     9  0.09           1.00

# You can use the wt function
count_pct(dt1, char1, wt = n)
#>     char1     n   pct pct_cumulative
#>    <char> <int> <num>          <num>
#> 1:      B    40  0.40           0.40
#> 2:      C    31  0.31           0.71
#> 3:      A    29  0.29           1.00