Simulates a day of work

This function sample from a data.frame or arrow connection all the trips needed to complete the a working day form taxi driver.

Usage

simulate_trips(
  arrow_con,
  start_datetime,
  start_zone,
  minutes_next_trip,
  end_datetime,
  valid_end_zones,
  closest_zone,
  borough_zones
)

Arguments

arrow_con

a data.frame or arrow connection with trips to sample from, with the columns described below.

start_datetime

a date with an associated time when the taxi driver start to work

start_zone

a number to select to the starting zone

minutes_next_trip

an integer to define the limit time needed before extending the filters

end_datetime

a date with an associated time when the taxi driver stop looking for new trips

valid_end_zones

a vector of number defining all possible zones to drive

closest_zone

a named vector pointing the closest zone from the taxi driver is waiting for a need trip in order to start a trip

borough_zones

a data.frame with the integer column LocationID with all possible zone ids and the list column id_list with contain all other zones ids related to the LocationID's borough

Value

A data.frame

Details

arrow_con must has the following columns

"Column Name"	"class"

year	integer

month	integer

PULocationID	integer

DOLocationID	integer

request_datetime	POSIXct POSIXt

dropoff_datetime	POSIXct POSIXt

driver_pay	numeric

tips	numeric

Examples

# Defining some zone context
valid_zones <- 1:3
closest_zones <- c(2L, 1L, 2L)
names(closest_zones) <- valid_zones
borough_zones <- data.frame(
  LocationID = valid_zones,
  id_list = I(list(valid_zones, valid_zones, valid_zones))
)

# Defining start point
start_datetime <- lubridate::make_datetime(2023L, 5L, 18L, 8L)
start_zone <- 1L

# Let's check trips from at 3 levels
# 1. Starting in the same zone
# 2. Starting in the closest zone
# 3. Starting in the farthest zone
valid_trips <- data.frame(
  year = 2023L,
  month = 5L,
  PULocationID = c(1L, 2L, 3L),
  DOLocationID = c(3L, 1L, 1L),
  request_datetime = c(
   start_datetime + lubridate::minutes(2L),
   start_datetime + lubridate::minutes(20L + 7L),
   start_datetime + lubridate::minutes(37L + 14L)
  ),
  dropoff_datetime = c(
   start_datetime + lubridate::minutes(20L),
   start_datetime + lubridate::minutes(37L),
   lubridate::make_datetime(2023L, 5L, 18L, 9L, 1L)
  ),
  driver_pay = 10,
  tips = 2 
)

simulate_trips(valid_trips,
               start_datetime = start_datetime,
               start_zone = start_zone,
               minutes_next_trip = 6L,
               end_datetime = start_datetime + lubridate::hours(1L),
               valid_end_zones = valid_zones,
               closest_zone = closest_zones,
               borough_zones = borough_zones)
#>   s_PULocationID s_DOLocationID  s_request_datetime  s_dropoff_datetime
#> 1              1              3 2023-05-18 08:02:00 2023-05-18 08:20:00
#> 2              2              1 2023-05-18 08:27:00 2023-05-18 08:37:00
#> 3              3              1 2023-05-18 08:51:00 2023-05-18 09:01:00
#>   s_driver_pay s_tips
#> 1           10      2
#> 2           10      2
#> 3           10      2